requirements.txt
asgiref==3.11.1
black==26.1.0
click==8.3.1
coverage==7.13.4
Django==6.0.2
django-cors-headers==4.9.0
django-ipware==7.0.1
django-structlog==10.0.0
djangorestframework==3.16.1
iniconfig==2.3.0
mypy_extensions==1.1.0
mysqlclient==2.2.8
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.2
python-ipware==3.0.0
pytokens==0.4.1
ruff==0.15.4
sqlparse==0.5.5
structlog==25.5.0
